YesNoModel updated using QuickType to auto generate a model using the JSON response of the API used in this app. Also, created toMessageEntity in YesNoModel to return a message based on the API response.
